+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unpublished
+### 🚀 Features
+- Added Bullets for Andes Message | Authors [Martin Becerra](https://github.com/mabecerra-MeLi)
+
 # 3.23.2
 ### 🛠 Bug fixes
 - Optional color var | Authors [joalonspint](https://github.com/joalonsopint)